requirements.txt
gradio==3.35.2
opencv-python==4.7.0.72
Pillow==9.4.0
requests
accelerate==0.20.3
diffusers==0.17.0
transformers==4.30.2
keras==2.10.0
tensorflow==2.10.1
xformers==0.0.20